[Feat]: PillManager 세부 구현, PillIController의 identifyPill 응답값 변경 (#5)

+++ clients/client-example/src/main/kotlin/org/mediscan/client/example/model/PillIdentificationResponseDto.kt
@@ -2,5 +2,5 @@ package org.mediscan.client.example.model;
 
 data class PillIdentificationResponseDto(
         var drugCode: String,
-        var confidence: Double
+        var confidence: Long
 )
